move pseudo terminal into vscode folder
