+++ extension/src/test/suite/vscode/pseudoTerminal.test.ts
@@ -3,8 +3,8 @@ import chai from 'chai'
 import sinonChai from 'sinon-chai'
 import { EventEmitter, Terminal, TerminalDataWriteEvent, window } from 'vscode'
 import { restore } from 'sinon'
-import { PseudoTerminal } from '../../PseudoTerminal'
-import { Disposable, Disposer } from '../../extension'
+import { Disposable, Disposer } from '../../../extension'
+import { PseudoTerminal } from '../../../vscode/pseudoTerminal'
 
 chai.use(sinonChai)
 const { expect } = chai
