+++ extension/src/vscode/pseudoTerminal.ts
